@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=1.22.4 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -29,23 +29,23 @@ dependencies:
       url: https://github.com/stringeecom/stringee_flutter_plugin.git
 
   #for ios
-  uuid: ^2.2.2
+  uuid: ^3.0.7
   callkeep:
     git:
       url: https://github.com/stringeecom/callkeep.git
       ref: master
 
   #for android
-  permission_handler: ^8.1.4+2
-  shared_preferences: ^2.0.6
-  firebase_core: ^2.4.0
-  firebase_messaging: ^14.2.0
-  device_info_plus: ^8.0.0
+  permission_handler: ^10.2.0
+  shared_preferences: ^2.0.18
+  firebase_core: ^2.7.0
+  firebase_messaging: ^14.2.5
+  device_info_plus: ^8.1.0
   flutter_local_notifications: ^13.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.5
 
 dev_dependencies:
   flutter_test:
