@@ -11,18 +11,22 @@ import 'package:ios_call_notification_sample/managers/ios_call_manager.dart';
 import 'package:stringee_flutter_plugin/stringee_flutter_plugin.dart';
 
 class SyncCall {
-  StringeeCall stringeeCall;
-  StringeeCall2 stringeeCall2;
+  StringeeCall? stringeeCall;
+  StringeeCall2? stringeeCall2;
 
-  int serial = 1;
-  String callId = '';
-  String uuid = ''; // uuid da su dung de show callkit
-  StringeeSignalingState callState = StringeeSignalingState.calling; // trang thai cua StringeeCall
+  int? serial = 1;
+  String? callId = '';
+  String? uuid = ''; // uuid da su dung de show callkit
+  StringeeSignalingState? callState =
+      StringeeSignalingState.calling; // trang thai cua StringeeCall
 
   bool userRejected = false;
-  bool userAnswered = false; // Người dùng đã click và nút answer ở màn hình incoming call của callkit hoặc của app
-  bool callAnswered = false; // StringeeCall đã được answer (đã gọi hàm answer của StringeeCall object)
-  bool audioSessionActived = false; // AudioSession của iOS đã được active thì khi answer call của Stringee mới kết nối thoại được
+  bool userAnswered =
+      false; // Người dùng đã click và nút answer ở màn hình incoming call của callkit hoặc của app
+  bool callAnswered =
+      false; // StringeeCall đã được answer (đã gọi hàm answer của StringeeCall object)
+  bool audioSessionActived =
+      false; // AudioSession của iOS đã được active thì khi answer call của Stringee mới kết nối thoại được
 
   bool endedCallkit = false;
   bool endedStringeeCall = false;
@@ -33,7 +37,8 @@ class SyncCall {
   bool hasLocalStream = false;
   bool hasRemoteStream = false;
   var _status = '';
-  bool mediaFirstTimeConnected = false; // Thêm biến này để check nếu là lần đầu media connected thì nếu là cuộc gọi video sẽ cho audio ra loa ngoài
+  bool mediaFirstTimeConnected =
+      false; // Thêm biến này để check nếu là lần đầu media connected thì nếu là cuộc gọi video sẽ cho audio ra loa ngoài
 
   set status(value) {
     _status = value;
@@ -43,16 +48,16 @@ class SyncCall {
   get status => _status;
 
   bool showedCallkit() {
-    return !uuid.isEmpty;
+    return uuid!.isNotEmpty;
   }
 
-  bool isThisCall(String callId, int serial) {
+  bool isThisCall(String? callId, int? serial) {
     return this.callId == callId && this.serial == serial;
   }
 
   bool ended() {
     bool callkitEndStatus = false;
-    if (uuid.isEmpty) {
+    if (uuid!.isEmpty) {
       callkitEndStatus = true;
     } else {
       callkitEndStatus = endedCallkit;
@@ -80,11 +85,11 @@ class SyncCall {
     * Hàm này thực hiện answer cuộc gọi đã show sử dụng CallKeep
     * Sau khi thành công sẽ nhận được sự kiện CallKeepPerformAnswerCallAction của CallKeep
     * **/
-    if (uuid == null || uuid.isEmpty) {
+    if (uuid == null || uuid!.isEmpty) {
       return;
     }
 
-    IOSCallManager.shared.callKeep.answerIncomingCall(uuid);
+    IOSCallManager.shared!.callKeep.answerIncomingCall(uuid!);
   }
 
   void answerIfConditionPassed() {
@@ -95,8 +100,12 @@ class SyncCall {
       3. Chua goi ham answer cua StringeeCall lan nao
       4. AudioSession da active
     **/
-    if (!userAnswered || callAnswered || !audioSessionActived || (stringeeCall == null && stringeeCall2 == null)) {
-      print('answerIfConditionPassed, condition has not been passed, userAnswered: ' +
+    if (!userAnswered ||
+        callAnswered ||
+        !audioSessionActived ||
+        (stringeeCall == null && stringeeCall2 == null)) {
+      print(
+          'answerIfConditionPassed, condition has not been passed, userAnswered: ' +
               userAnswered.toString() +
               ", callAnswered: " +
               callAnswered.toString() +
@@ -106,12 +115,13 @@ class SyncCall {
     }
 
     // Cập nhật giao diện từ incomingCall ==> calling
-    if (IOSCallManager.shared.callScreenKey != null && IOSCallManager.shared.callScreenKey.currentState != null) {
-      IOSCallManager.shared.callScreenKey.currentState.changeToCallingUI();
+    if (IOSCallManager.shared!.callScreenKey != null &&
+        IOSCallManager.shared!.callScreenKey!.currentState != null) {
+      IOSCallManager.shared!.callScreenKey!.currentState!.changeToCallingUI();
     }
 
     if (stringeeCall != null) {
-      stringeeCall.answer().then((result) {
+      stringeeCall!.answer().then((result) {
         String message = result['message'];
         bool status = result['status'];
         print("answer: " + message);
@@ -121,7 +131,7 @@ class SyncCall {
         }
       });
     } else if (stringeeCall2 != null) {
-      stringeeCall2.answer().then((result) {
+      stringeeCall2!.answer().then((result) {
         String message = result['message'];
         bool status = result['status'];
         print("answer: " + message);
@@ -131,26 +141,28 @@ class SyncCall {
         }
       });
     }
-
   }
 
-  Future<bool> reject() async {
+  Future<bool?> reject() async {
     if (stringeeCall == null && stringeeCall2 == null) {
-      print("SyncCall reject failed, call: " + stringeeCall.toString() + ", call2: " + stringeeCall2.toString());
+      print("SyncCall reject failed, call: " +
+          stringeeCall.toString() +
+          ", call2: " +
+          stringeeCall2.toString());
       return false;
     }
 
-    bool status;
+    bool? status;
 
     if (stringeeCall != null) {
-      await stringeeCall.reject().then((result) {
+      await stringeeCall!.reject().then((result) {
         String message = result['message'];
         status = result['status'];
         print("SyncCall reject, message: " + message);
         endCallIfNeed();
       });
     } else {
-      await stringeeCall2.reject().then((result) {
+      await stringeeCall2!.reject().then((result) {
         String message = result['message'];
         status = result['status'];
         print("SyncCall reject, message: " + message);
@@ -161,26 +173,35 @@ class SyncCall {
     return status;
   }
 
-  Future<bool> hangup() async {
+  Future<bool?> hangup() async {
     if (stringeeCall == null && stringeeCall2 == null) {
-      print("SyncCall hangup failed, call: " + stringeeCall.toString() + ", call2: " + stringeeCall2.toString());
+      print("SyncCall hangup failed, call: " +
+          stringeeCall.toString() +
+          ", call2: " +
+          stringeeCall2.toString());
       return false;
     }
 
-    bool status;
+    bool? status;
 
     if (stringeeCall != null) {
-      await stringeeCall.hangup().then((result) {
+      await stringeeCall!.hangup().then((result) {
         String message = result['message'];
         status = result['status'];
-        print("SyncCall hangup, message: " + message + ", status: " + status.toString());
+        print("SyncCall hangup, message: " +
+            message +
+            ", status: " +
+            status.toString());
         endCallIfNeed();
       });
     } else {
-      await stringeeCall2.hangup().then((result) {
+      await stringeeCall2!.hangup().then((result) {
         String message = result['message'];
         status = result['status'];
-        print("SyncCall hangup, message: " + message + ", status: " + status.toString());
+        print("SyncCall hangup, message: " +
+            message +
+            ", status: " +
+            status.toString());
         endCallIfNeed();
       });
     }
@@ -188,9 +209,12 @@ class SyncCall {
     return status;
   }
 
-  void mute({bool isMute}) {
+  void mute({bool? isMute}) {
     if (stringeeCall == null && stringeeCall2 == null) {
-      print("SyncCall mute failed, call: " + stringeeCall.toString() + ", call2: " + stringeeCall2.toString());
+      print("SyncCall mute failed, call: " +
+          stringeeCall.toString() +
+          ", call2: " +
+          stringeeCall2.toString());
       return;
     }
 
@@ -201,24 +225,27 @@ class SyncCall {
     }
 
     if (stringeeCall != null) {
-      stringeeCall.mute(this.isMute);
+      stringeeCall!.mute(this.isMute);
     } else {
-      stringeeCall2.mute(this.isMute);
+      stringeeCall2!.mute(this.isMute);
     }
     updateUI();
   }
 
   Future<bool> setSpeakerphoneOn() async {
     if (stringeeCall == null && stringeeCall2 == null) {
-      print("SyncCall setSpeakerphoneOn failed, call: " + stringeeCall.toString() + ", call2: " + stringeeCall2.toString());
+      print("SyncCall setSpeakerphoneOn failed, call: " +
+          stringeeCall.toString() +
+          ", call2: " +
+          stringeeCall2.toString());
       return false;
     }
 
     isSpeaker = !isSpeaker;
     if (stringeeCall != null) {
-      await stringeeCall.setSpeakerphoneOn(isSpeaker);
+      await stringeeCall!.setSpeakerphoneOn(isSpeaker);
     } else {
-      await stringeeCall2.setSpeakerphoneOn(isSpeaker);
+      await stringeeCall2!.setSpeakerphoneOn(isSpeaker);
     }
     updateUI();
     return true;
@@ -226,28 +253,34 @@ class SyncCall {
 
   void switchCamera() {
     if (stringeeCall == null && stringeeCall2 == null) {
-      print("SyncCall switchCamera failed, call: " + stringeeCall.toString() + ", call2: " + stringeeCall2.toString());
+      print("SyncCall switchCamera failed, call: " +
+          stringeeCall.toString() +
+          ", call2: " +
+          stringeeCall2.toString());
       return;
     }
 
     if (stringeeCall != null) {
-      stringeeCall.switchCamera();
+      stringeeCall!.switchCamera();
     } else {
-      stringeeCall2.switchCamera();
+      stringeeCall2!.switchCamera();
     }
   }
 
   Future<bool> enableVideo() async {
     if (stringeeCall == null && stringeeCall2 == null) {
-      print("SyncCall enableVideo failed, call: " + stringeeCall.toString() + ", call2: " + stringeeCall2.toString());
+      print("SyncCall enableVideo failed, call: " +
+          stringeeCall.toString() +
+          ", call2: " +
+          stringeeCall2.toString());
       return false;
     }
 
     videoEnabled = !videoEnabled;
     if (stringeeCall != null) {
-      await stringeeCall.enableVideo(videoEnabled);
+      await stringeeCall!.enableVideo(videoEnabled);
     } else {
-      await stringeeCall2.enableVideo(videoEnabled);
+      await stringeeCall2!.enableVideo(videoEnabled);
     }
     updateUI();
     return true;
@@ -258,7 +291,8 @@ class SyncCall {
       return;
     }
 
-    if ((stringeeCall != null && stringeeCall.isVideoCall) || (stringeeCall2 != null && stringeeCall2.isVideoCall)) {
+    if ((stringeeCall != null && stringeeCall!.isVideoCall) ||
+        (stringeeCall2 != null && stringeeCall2!.isVideoCall)) {
       mediaFirstTimeConnected = true;
       isSpeaker = false;
       setSpeakerphoneOn();
@@ -266,12 +300,13 @@ class SyncCall {
   }
 
   void endCallIfNeed() {
-    IOSCallManager.shared.clearDataEndDismiss();
+    IOSCallManager.shared!.clearDataEndDismiss();
   }
 
   void updateUI() {
-    if (IOSCallManager.shared.callScreenKey != null && IOSCallManager.shared.callScreenKey.currentState != null) {
-      IOSCallManager.shared.callScreenKey.currentState.setState(() {});
+    if (IOSCallManager.shared!.callScreenKey != null &&
+        IOSCallManager.shared!.callScreenKey!.currentState != null) {
+      IOSCallManager.shared!.callScreenKey!.currentState!.setState(() {});
     }
   }
 
@@ -283,38 +318,37 @@ class SyncCall {
 
   bool isVideoCall() {
     if (stringeeCall != null) {
-      return stringeeCall.isVideoCall;
+      return stringeeCall!.isVideoCall;
     } else if (stringeeCall2 != null) {
-      return stringeeCall2.isVideoCall;
+      return stringeeCall2!.isVideoCall;
     }
 
     return false;
   }
 
-  String from() {
+  String? from() {
     if (stringeeCall != null) {
-      return stringeeCall.from;
+      return stringeeCall!.from;
     } else if (stringeeCall2 != null) {
-      return stringeeCall2.from;
+      return stringeeCall2!.from;
     }
     return "";
   }
 
-  String to() {
+  String? to() {
     if (stringeeCall != null) {
-      return stringeeCall.to;
+      return stringeeCall!.to;
     } else if (stringeeCall2 != null) {
-      return stringeeCall2.to;
+      return stringeeCall2!.to;
     }
     return "";
   }
 
   void destroy() {
     if (stringeeCall != null) {
-      return stringeeCall.destroy();
+      return stringeeCall!.destroy();
     } else if (stringeeCall2 != null) {
-      stringeeCall2.destroy();
+      stringeeCall2!.destroy();
     }
   }
-
 }
